symlink fixup
First I promised that we'd change as little as possible in mednafen to ease upstream integration, and now I check in 1730 changes in one go.  What happened?  Symlinks happened.

When I first committed this repo, I looped the original tarfile through a `tar -xf`, `tar -chf` pair to expand out all symlinks.  This was intentional; symlinks are a very developer-hostile feature that make understanding directory structure difficult, and have technical problems on many systems.  But that wasn't enough; it's too much stuff and too hard to work with.  Mednafen uses way too many symlinked files, and you have to remember to edit both versions of a file as well as grok the various include paths, because files can and do use both the original notion and the symlinked notion of where their dependencies are.

With this commit, I remove the symlinked versions of files in include/mednafen/*, and global replace #include <mednafen/* with #include <src/* so that files consistently point to the right thing.
There are probably a few oddball files this doesn't work right on, but nothing in the parts that I'm compiling right now so tweaks will have to wait for later.

To take later updates, we have to:
1) Download the new mednafen tarball
2) Untar it to a filesystem that fully supports developer-hostile and confusing features.
3) Retar it with -h to eliminate the the developer-hostile and confusing features.
4) Clear out the git working directory.
5) Untar the file from #3 into the git working directory.
6) Global replace `#include <mednafen/` with `#include <src/`
7) Commit.

Ugggg.