Rework how VideoFormatChanged / SoundFormatChanged work
Instead of being lazily handled inside the cores, there's now a separate FormatsChanged function that the frontend must call when the formats change.  This is a bit intrusive in the code, unfortunately, but it's less bad than the old alloc_invis everywhere.  This can work better with waterbox memory semantics because we can just call FormatsChanged() once before sealing.